requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
psycopg2==2.9.7
python-dotenv==1.0.0
sqlparse==0.4.4
daphne==4.1.2
channels==4.1.0
channels-redis==4.2.0
pillow==10.4.0
django-ratelimit==4.1.0
django-import-export==4.1.1
django-redis==5.4.0
requests==2.32.3